requirements.txt
streamlit==1.37.0
pandas==2.2.2
pymupdf==1.24.9
pytesseract==0.3.10
Pillow==10.4.0
matplotlib==3.9.2
together==1.2.1
openpyxl==3.1.5
python-docx==1.1.2
python-dotenv==1.0.1